+++ NEWS.md
@@ -5,6 +5,7 @@
 * Add `gcs_first` and `gcs_last` to autosave your file workspace to GCS
 * Better message feedback on file sizes
 * Add `gcs_save_all` and `gcs_load_all` which will zip, save/load and upload/download a directory
+* Add use of `_gcssave.yaml` file to control `gcs_first/last` behaviour
 
 # 0.3.0